+++ epstein_files/output/output.py
@@ -24,7 +24,7 @@ from epstein_files.output.rich import *
 from epstein_files.output.site.internal_links import (AUTHORS_USING_SIGNATURES, FILES_THAT_ARE_NEITHER_EMAILS_NOR,
      HIS_EMAILS, HIS_TEXT_MESSAGES, SELECTIONS_FROM)
 from epstein_files.output.site.sites import EMAILERS_TABLE_PNG_PATH, HtmlDir
-from epstein_files.output.title_page import print_other_page_link, section_header
+from epstein_files.output.title_page import section_header
 from epstein_files.people.interesting_people import EMAILERS_TO_PRINT
 from epstein_files.people.names import *
 from epstein_files.people.person import Person
@@ -148,9 +148,9 @@ def print_emails_section(epstein_files: EpsteinFiles, printer: DocPrinter) -> No
     Print emails grouped by participant with summary tables.
     Emails can be printed more than once (e.g. in the sender's section and each recipient's).
     """
-    printer.print_section_subtitle((SELECTIONS_FROM if not args.all_emails else '') + HIS_EMAILS)
+    printer.print_section_subtitle((SELECTIONS_FROM if not args.all_emailers else '') + HIS_EMAILS)
     all_emailers = sorted(epstein_files.emailers, key=lambda person: person.earliest_email_at)
-    emailers_to_print = all_emailers if args.all_emails else epstein_files.person_objs(EMAILERS_TO_PRINT)
+    emailers_to_print = all_emailers if args.all_emailers else epstein_files.person_objs(EMAILERS_TO_PRINT)
     printer.print(_section_summary_table(Person.emailer_info_table(emailers_to_print)))
     num_since_color_key = 0
 
@@ -182,7 +182,7 @@ def print_emails_section(epstein_files: EpsteinFiles, printer: DocPrinter) -> No
     log_msg = f"Rewrote {len(Email.rewritten_header_ids)} of {len(printer.printed_emails)} email headers"
     logger.warning(f"  -> {log_msg}, {len(fwded_articles)} of the Emails printed were forwarded articles.")
 
-    if args.all_emails:
+    if args.all_emailers:
         _verify_all_emails_were_printed(epstein_files, printer.printed_emails)
 
 
@@ -253,7 +253,6 @@ def print_other_files_section(epstein_files: EpsteinFiles, printer: DocPrinter) 
     title_pfx = '' if args.all_other_files else 'Selected '
     category_table = OtherFile.summary_table(files, title_pfx=title_pfx)
     printer.print_section_subtitle(f"{FIRST_FEW_LINES} of {len(files)} {title_pfx}{FILES_THAT_ARE_NEITHER_EMAILS_NOR}")
-    print_other_page_link(epstein_files)  # TODO: not in custom HTML
     printer.print(_section_summary_table(category_table))
 
     # If --all-other-files is enables, print the biographical panels, otherwise just print a big table