+++ epstein_files/output/title_page.py
@@ -36,6 +36,7 @@ OTHER_PAGE_MSG_STYLE = 'gray78 dim'
 STR_VAL_STYLE = 'cornsilk1 italic'
 STR_VAL_STYLE_ALT = 'light_yellow3 italic'
 SECTION_HEADER_STYLE = 'bold black on color(146)'
+TITLE_STYLE = 'black on grey23'  # color(103)'
 
 DATASET_MSG_LINES = [
     f"This dataset includes everything from the {HOUSE_OVERSIGHT_TRANCHE}",
@@ -55,6 +56,7 @@ COLOR_KEYS = [
 
 
 def color_key() -> Padding:
+    """Generate the color highlight color->label panel."""
     color_table = build_table('Rough Guide to Highlighted Colors', show_header=False)
     num_colors = len(COLOR_KEYS)
     row_number = 0
@@ -70,57 +72,32 @@ def color_key() -> Padding:
     return vertically_pad(color_table)
 
 
-def print_color_key() -> None:
-    color_table = build_table('Rough Guide to Highlighted Colors', show_header=False)
-    num_colors = len(COLOR_KEYS)
-    row_number = 0
+def header_elements() -> list[RenderableType]:
+    title = Text('', justify='center').append('The Epstein Files', style='underline bold')
+    title_panel = Panel(title, box=box.DOUBLE_EDGE, expand=True, padding=(2, 2), style=TITLE_STYLE, width=TITLE_WIDTH)
+    title_with_vertical_margin = vertically_pad(title_panel)
 
-    for i in range(0, site_config.num_color_key_cols):
-        color_table.add_column(f"color_col_{i}", justify='center')
+    elements = [
+        MOBILE_WARNING_TXT,
+        title_with_vertical_margin,
+        SUBSTACK_POST_TXT_MSGS_LINK,
+    ]
 
-    while (row_number * site_config.num_color_key_cols) < num_colors:
-        idx = row_number * site_config.num_color_key_cols
-        color_table.add_row(*COLOR_KEYS[idx:(idx + site_config.num_color_key_cols)])
-        row_number += 1
+    if not args.mobile:
+        substack_link = SUBSTACK_POST_TXT_MSGS_LINK.url_link
+        substack_link.stylize(f'{SUBSTACK_POST_LINK_STYLE} dim')
+        elements.append(substack_link)
 
-    print_centered(vertically_pad(color_table))
-
-
-def print_other_page_link(epstein_files: 'EpsteinFiles') -> None:
-    if args._site == Site.CURATED_MOBILE:
-        return
-    elif args._site == Site.CURATED:
-        txt = THE_OTHER_PAGE_TXT + Text(f' is uncurated and has all {len(epstein_files.emails):,} emails')
-        txt.append(f" and {len(epstein_files.other_files):,} unclassifiable files")
-    else:
-        txt = THE_OTHER_PAGE_TXT + (f' displays a curated collection of emails and')
-        txt.append(" unclassifiable files of particular interest")
-
-    print_centered(parenthesize(txt), style=OTHER_PAGE_MSG_STYLE)
-    chrono_emails_link = link_text_obj(Site.get_url(Site.EMAILS_CHRONOLOGICAL), 'a page', style='light_slate_grey bold')
-    chrono_emails_txt = Text(f"there's also ").append(chrono_emails_link)
-    chrono_emails_txt.append(' with all the emails in chronological order')
-    print_centered(parenthesize(chrono_emails_txt), style=OTHER_PAGE_MSG_STYLE)
+    elements.append(join_texts(CRYPTADAMUS_SOCIAL_LINKS, join=site_config.social_link_separator))
+    return elements
 
 
+# TODO: move to .rich
 def section_header(msg: str, style: str = SECTION_HEADER_STYLE, is_centered: bool = False) -> Padding:
     """Make a padded Panel that's centered."""
-    if args._site == Site.CURATED:
-        console.line(2)
-        print_section_links()
-
     panel = Panel(Text(msg, justify='center'), expand=True, padding=(1, 1), style=style)
     panel = Align.center(panel) if is_centered else panel
     return Padding(panel, site_config.section_header_padding)
-
-
-def title_page_top_elements() -> list[RenderableType]:
-    return [
-        *_header_elements(),
-        Site.build_directory(),
-        starred_header_txt(site_config.not_all_files_warning, num_spaces=0, num_stars=0),
-        *DATASET_MSG_TXTS,
-    ]
 
 
 def title_page_bottom_elements(epstein_files: 'EpsteinFiles') -> list[RenderableType]:
@@ -157,27 +134,3 @@ def _abbreviations_table() -> Padding:
             table.add_row(highlighter(k), v)
 
     return vertically_pad(table)
-
-
-def _header_elements() -> list[RenderableType]:
-    title = Text('', justify='center').append('The Epstein Files', style='underline bold')
-    title_panel = Panel(title, box=box.DOUBLE_EDGE, expand=True, padding=(2, 2), style=TITLE_STYLE, width=TITLE_WIDTH)
-    title_with_vertical_margin = vertically_pad(title_panel)
-
-    # unwrapped, css_props = unwrap_rich(title_with_vertical_margin)
-    # print(f"rich obj {unwrapped} has unwrapped title_panel css: {css_props}")
-    # import pdb;pdb.set_trace()
-
-    elements = [
-        MOBILE_WARNING_TXT,
-        title_with_vertical_margin,
-        SUBSTACK_POST_TXT_MSGS_LINK,
-    ]
-
-    if not args.mobile:
-        substack_link = SUBSTACK_POST_TXT_MSGS_LINK.url_link
-        substack_link.stylize(f'{SUBSTACK_POST_LINK_STYLE} dim')
-        elements.append(substack_link)
-
-    elements.append(join_texts(CRYPTADAMUS_SOCIAL_LINKS, join=site_config.social_link_separator))
-    return elements