+++ scripts/build_pages.sh
@@ -9,31 +9,31 @@ set -e
 THIS_DIR=$(dirname -- "$(readlink -f -- "$0";)";)
 source "$THIS_DIR/bash_lib/shared.sh"
 
+BUILD_DIR=${1:-docs/}
+GENERATE_CMD="$GENERATE_CMD --build-dir $BUILD_DIR"
+GENERATE_MOBILE_CMD="$GENERATE_MOBILE_CMD --build-dir $BUILD_DIR"
+GENERATE_SIDE_PANELS_CMD="$GENERATE_CMD --side-panel-notes"
+
 CATEGORIES=(
     crypto
     girls
     money
 )
 
-BUILD_DIR=${1:-docs/}
-GENERATE_CMD="$GENERATE_CMD --build-dir $BUILD_DIR"
-GENERATE_SIDE_PANELS_CMD="$GENERATE_CMD --side-panel-notes"
-print_deploy_step "Building pages to BUILD_DIR '$BUILD_DIR'"
-
 
 if [[ -z $SKIP_CHRONO ]]; then
-    print_deploy_step "Building most interesting page with $GENERATE_SIDE_PANELS_CMD --output-most-interesting"
+    print_deploy_step "Building" "--output-most-interesting"
     $GENERATE_SIDE_PANELS_CMD --output-most-interesting
-    print_deploy_step "Building curated chronological page with '$GENERATE_SIDE_PANELS_CMD'"
+    print_deploy_step "Building" "--output-chrono"
     $GENERATE_SIDE_PANELS_CMD --output-chrono
-    print_deploy_step "Building curated chronological mobile page..."
+    print_deploy_step "Building" "--output-chrono --mobile"
     $GENERATE_MOBILE_CMD --output-chrono
 else
     print_deploy_step "Skipping chronological builds..."
 fi
 
 # Fast pages
-print_deploy_step "Building notes pages..."
+print_deploy_step "Building" "--output-notes"
 $GENERATE_CMD --output-notes
 print_deploy_step "Building biographies page..."
 $GENERATE_CMD --output-bios
@@ -61,11 +61,11 @@ else
     print_deploy_step "Building other files table page..."
     $GENERATE_CMD --all-other-files
     print_deploy_step "Building curated page..."
-    $GENERATE_SIDE_PANELS_CMD
+    $GENERATE_SIDE_PANELS_CMD --output-curated
     print_deploy_step "Building curated mobile page... "
-    $GENERATE_MOBILE_CMD
+    $GENERATE_MOBILE_CMD --output-curated
     print_deploy_step "Building all emails page..."
-    $GENERATE_CMD --all-emails
+    $GENERATE_CMD --all-emailers
     print_deploy_step "Building emails chronological page..."
     $GENERATE_SIDE_PANELS_CMD --all-emails-chrono
 fi